Added check on delete recipes to check if user owns or is admin before delete

+++ cookbook/routes.py
@@ -246,9 +246,14 @@ def delete_recipe(recipe_id):
     """
     Deletes recipe clicked on by user from the recipes page
     """
-    recipe = Recipes.query.get_or_404(recipe_id)
-    db.session.delete(recipe)
-    db.session.commit()
+    find_recipe = Recipes.query.get_or_404(recipe_id)
+    print(find_recipe)
+    if current_user.is_admin == True or current_user.id == find_recipe.owner:
+        recipe = Recipes.query.get_or_404(recipe_id)
+        db.session.delete(recipe)
+        db.session.commit()
+    else:
+        print('You can not delete this')
     return redirect(url_for('routes.recipes'))
